Fix the problem of storing redundant sessions
In the original code, when the LoginGet() is called and the arguments of the signature and the callback url are not empty, it stores the same session in the database. Besides, when the LoginPost() wiht the non-empty signature and the callback url, the session will be created and then saved.