[OWL-479] Agent adds exe file permission to plugin before it runs
Originally, plugin files easily get stuck because of exe
permission problems. Alleviate this problem in this commit by
detecting if the plugin files are scripts with shebang.

If the plugin files are scripts with shebang and having no exe
permission, just call the interpreter(python, bash) to execute
the scripts.

Attached test directory are the files used for unit test. The
content in script are not important.