[OWL-1730][common][nqm-mng] Recover panics
Since a panic in a goroutine causes the stop of the whole process, I
added the reovery mechanism in `service/queue/`. In this commit, I also
improved the mechanism. If it drains nothing from the queue, it doesn't
call the method in the data layer. For the debug messages, I implemented
`String()` method of `NqmAgentHeartbeatRequest`. I made `nqm.Agent` use
`JsonTime` instead of `time.Time` and relative changes are added.