[OWL-1730][nqm-mng] Remove `mutex`
Previously we used `Mutex` to control the status of the queue service.
It is actually unnecessary becuase we are very sure that in our
application, the `Start()` is called first, then `Put()` or `Stop()` is
called. Moreover, I use `time.Sleep()` instead to make all the
`Enqueue()` be done before flushing.